Id/Idref Namespace support (read-only)
id & idref attributes are QNames, so the associated xmlns entries need
to be resolved properly.

This patches all STIX top-level objects & STIXPackage to read the
namespace info from the parser nsmap, and make it available with
id/idref as a pair.

The api objects from_obj() is modified to retrieve the pair and the
namespace is available as id_ns & idref_ns.

Read-only -- no write support.