Optional idgen namespace gathering in to_xml()
Since the idgen namespace settings are at the module-level, they could
cause thread-safety issues, depending on the use.

This patch moves the idgen gathering up to the to_xml() and out of
nsparser, and optional by using the include_idgen=T/F parameter to
to_xml.

To implement a thread-safe to_xml() call, the caller should disable the
include_idgen, and provide the needed namespaces using the ns_dict
parameter.